2025-11-06 21h55 I almost forgot to provide reference verbs that retrieve the first superior relationships that a column participates in if any. They exist now: Value.FirstSuperiorRelationship, Value.firstSuperiorCandidateRelationship, and Value.firstSuperiorNonCandidateRelationship.  The general FirstSuperiorRelationship tries to get the Candidate version first, and if no such exists, retrieves the non-candidate version (i.e. it prefers to stay in-schema.)
The idea is for the rules to check if a column is in a superior relationship, and not produce an @Column for it in that case, but instead count on the
relationship being established and using the relationship to resolve the getter for the attribute, while the setter throws a "NotSupported" exception.
This will allow the HKey's to get the values from the JPA entities needed to populate the history with _columns_ instead of relationships (trying to
track historical relationships is beyond my grasp or fathoming. Is a generic relationship to history required? To a specific version in effect at the same time?
Just too many questions, and most of them requiring rather complex answers.)

Only the SuperiorCandidateRelationships for a Table are going to be specified at first; have to start somewhere, and the plan is to resolve the non-candidate
relations in the Obj layer instead, where cross-schema operations will be supported automagically and a consolidated view of them produced.  I've realized I don't
like the way I handle the class codes outside a schema; they need to be remapped on the fly by the Obj layer to consistent, predictable, _catalogued_ values that a
client application can retrieve easily and use to map its data against a generic server that integrates that particular schema's application specifications and
interface profiles.

Remember, the goal is to manufacture integrated application frameworks that can still function with the "normal" client for that particular set of services, but also
support the other applications that have been integrated into the server. As there is no way to coordinate 6-character hex codes that have a first alphabetic character
across _all_ models defined by people, I need a way to _map_ their class codes portably.